schannel SSL: process extra data buffer immediately
This makes it so that when there is an extra data buffer ready to be
processed during the handshake it is processed before returning from
schannel_connect_step2.  Without this change I've seen WinCE hang
when schannel_connect_step2 returns and calls Curl_socket_ready.

config-win32ce.h - process.h is not available on WinCE.